refactor(users): some adjustments for mfa
